Move custom hive catalog to hivelink-core
