+++ build.gradle
@@ -495,7 +495,7 @@ project(':iceberg-hive-metastore') {
   }
 }
 
-project(':iceberg-hivelink') {
+project(':iceberg-hivelink-core') {
   dependencies {
     compile project(':iceberg-hive-metastore')
 
@@ -872,7 +872,7 @@ if (jdkVersion == '8') {
       compile project(':iceberg-arrow')
       compile project(':iceberg-hive-metastore')
       compile project(':iceberg-spark')
-      compile project(':iceberg-hivelink')
+      compile project(':iceberg-hivelink-core')
 
       compileOnly "org.apache.avro:avro"
       compileOnly("org.apache.spark:spark-hive_2.11") {
@@ -982,7 +982,7 @@ project(':iceberg-spark3') {
     compile project(':iceberg-arrow')
     compile project(':iceberg-hive-metastore')
     compile project(':iceberg-spark')
-    compile project(':iceberg-hivelink')
+    compile project(':iceberg-hivelink-core')
 
     compileOnly "org.apache.avro:avro"
     compileOnly("org.apache.spark:spark-hive_2.12") {
