--- hive-metastore/src/main/java/org/apache/iceberg/hive/HiveMetadataPreservingCatalog.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iceberg.hive;
-
-import org.apache.hadoop.conf.Configuration;
-import org.apache.iceberg.TableOperations;
-import org.apache.iceberg.catalog.TableIdentifier;
-import org.apache.iceberg.hadoop.HadoopFileIO;
-
-
-/**
- * A {@link HiveCatalog} which uses {@link HiveMetadataPreservingTableOperations} underneath.
- *
- * This catalog should be only be used by metadata publishers wanting to publish/update Iceberg metadata to an existing
- * Hive table while preserving the current Hive metadata
- */
-public class HiveMetadataPreservingCatalog extends HiveCatalog {
-
-  public HiveMetadataPreservingCatalog(Configuration conf) {
-    super(conf);
-  }
-
-  @Override
-  public TableOperations newTableOps(TableIdentifier tableIdentifier) {
-    String dbName = tableIdentifier.namespace().level(0);
-    String tableName = tableIdentifier.name();
-    return new HiveMetadataPreservingTableOperations(conf(), clientPool(), new HadoopFileIO(conf()), name(), dbName,
-        tableName);
-  }
-}
