+++ hivelink-core/src/test/resources/hive-schema-3.1.0.derby.sql
