+++ settings.gradle
@@ -37,7 +37,7 @@ include 'spark3-extensions'
 include 'spark3-runtime'
 include 'pig'
 include 'hive-metastore'
-include 'hivelink'
+include 'hivelink-core'
 include 'nessie'
 include 'runtime'
 
@@ -60,7 +60,7 @@ project(':spark3-extensions').name = 'iceberg-spark3-extensions'
 project(':spark3-runtime').name = 'iceberg-spark3-runtime'
 project(':pig').name = 'iceberg-pig'
 project(':hive-metastore').name = 'iceberg-hive-metastore'
-project(':hivelink').name = 'iceberg-hivelink'
+project(':hivelink-core').name = 'iceberg-hivelink-core'
 project(':nessie').name = 'iceberg-nessie'
 project(':runtime').name = 'iceberg-runtime'
